Adds file output to `results` function
Adds multiple lines to allow for the printing of results to be added to a file called `log_results.txt`. Plenty of comments are added to explain the implementation and to cite the sources for where I learned how to use the modules I imported, or for more documentation on things I did.